# libs/openchallenges/api-description/openapi/challenge.openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: OpenChallenges Challenge API
  description: Discover, explore, and contribute to open biomedical challenges.
  license:
    name: Apache 2.0
    url: https://github.com/Sage-Bionetworks/sage-monorepo/blob/main/LICENSE.txt
  contact:
    name: Support
    url: https://github.com/Sage-Bionetworks/sage-monorepo
  x-logo:
    url: https://dev.openchallenges.io/img/unsafe/logo/OpenChallenges-logo.png
servers:
  - url: http://localhost/v1
...
paths:
...
  /challenges/{challengeId}/contributions:
    parameters:
      - $ref: '#/components/parameters/challengeId'
    get:
      tags:
        - ChallengeContribution
      summary: List challenge contributions
      description: List challenge contributions
      operationId: listChallengeContributions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeContributionsPage'
          description: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - ChallengeContribution
      summary: Create a new contribution for a challenge
      description: |
        Creates a new contribution record associated with a challenge ID.
      operationId: addChallengeContribution
      security:
        - apiBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeContributionCreateRequest'
      responses:
        '201':
          description: Contribution created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeContribution'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - ChallengeContribution
      summary: Delete all contributions for a specific challenge
      description: |
        Deletes all associated contributions for a given challenge, identified by its ID.
        This action is irreversible.
      operationId: deleteAllChallengeContributions
      security:
        - apiBearerAuth: []
      responses:
        '204':
          description: Deletion successful
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /challenges/{challengeId}/contributions/{challengeContributionId}:
    parameters:
      - $ref: '#/components/parameters/challengeId'
      - $ref: '#/components/parameters/challengeContributionId'
    get:
      tags:
        - ChallengeContribution
      summary: Get a specific challenge contribution
      description: |
        Retrieves a specific contribution record for a challenge, identified by its ID.
      operationId: getChallengeContribution
      responses:
        '200':
          description: Challenge contribution retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeContribution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - ChallengeContribution
      summary: Update an existing challenge contribution
      description: |
        Updates an existing contribution record for a challenge.
        Only the organization ID and role can be modified.
      operationId: updateChallengeContribution
      security:
        - apiBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeContributionUpdateRequest'
      responses:
        '200':
          description: Contribution updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeContribution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - ChallengeContribution
      summary: Delete a specific challenge contribution
      description: |
        Deletes a specific contribution record for a challenge, identified by its ID.
        This action is irreversible.
      operationId: deleteChallengeContribution
      security:
        - apiBearerAuth: []
      responses:
        '204':
          description: Contribution deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    apiBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: api_key
      description: API key obtained from /auth/login endpoint
  schemas:
...
    OrganizationId:
      description: The unique identifier of an organization
      type: integer
      format: int64
      example: 1
...
    PageMetadata:
      type: object
      description: The metadata of a page.
      properties:
        number:
          description: The page number.
          type: integer
          format: int32
          example: 99
        size:
          description: The number of items in a single page.
          type: integer
          format: int32
          example: 99
        totalElements:
          description: Total number of elements in the result set.
          type: integer
          format: int64
          example: 99
        totalPages:
          description: Total number of pages in the result set.
          type: integer
          format: int32
          example: 99
        hasNext:
          description: Returns if there is a next page.
          type: boolean
          example: true
        hasPrevious:
          description: Returns if there is a previous page.
          type: boolean
          example: true
      required:
        - number
        - size
        - totalElements
        - totalPages
        - hasNext
        - hasPrevious
    ChallengeId:
      description: The unique identifier of the challenge.
      type: integer
      format: int64
      example: 1
...
    BasicError:
      type: object
      description: Problem details (tools.ietf.org/html/rfc7807)
      properties:
        title:
          type: string
          description: A human readable documentation for the problem type
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human readable explanation specific to this occurrence of the problem
        type:
          type: string
          description: An absolute URI that identifies the problem type
      required:
        - title
        - status
      x-java-class-annotations:
        - '@lombok.AllArgsConstructor'
        - '@lombok.Builder'
...
    ChallengeContributionId:
      description: The unique identifier of a challenge contribution
      type: integer
      format: int64
      example: 1
    ChallengeContributionRole:
      description: The nature of a challenge contribution.
      type: string
      enum:
        - challenge_organizer
        - data_contributor
        - sponsor
      example: challenge_organizer
    ChallengeContribution:
      type: object
      description: A challenge contribution.
      properties:
        id:
          $ref: '#/components/schemas/ChallengeContributionId'
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
        organizationId:
          $ref: '#/components/schemas/OrganizationId'
        role:
          $ref: '#/components/schemas/ChallengeContributionRole'
      required:
        - id
        - challengeId
        - organizationId
        - role
    ChallengeContributionsPage:
      type: object
      description: A page of challenge challenge contributions.
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            challengeContributions:
              description: A list of challenge contributions.
              type: array
              items:
                $ref: '#/components/schemas/ChallengeContribution'
          required:
            - challengeContributions
      x-java-class-annotations:
        - '@lombok.Builder'
    ChallengeContributionCreateRequest:
      type: object
      description: A challenge contribution to be created.
      properties:
        organizationId:
          $ref: '#/components/schemas/OrganizationId'
        role:
          $ref: '#/components/schemas/ChallengeContributionRole'
      required:
        - organizationId
        - role
    ChallengeContributionUpdateRequest:
      type: object
      description: A challenge contribution update request.
      properties:
        organizationId:
          $ref: '#/components/schemas/OrganizationId'
        role:
          $ref: '#/components/schemas/ChallengeContributionRole'
      required:
        - organizationId
        - role
...
  parameters:
...
    challengeId:
      name: challengeId
      in: path
      description: The unique identifier of the challenge.
      required: true
      schema:
        $ref: '#/components/schemas/ChallengeId'
    challengeContributionId:
      name: challengeContributionId
      in: path
      description: The unique identifier of a challenge contribution
      required: true
      schema:
        $ref: '#/components/schemas/ChallengeContributionId'
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    InternalServerError:
      description: The request cannot be fulfilled due to an unexpected server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    NotFound:
      description: The specified resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    Forbidden:
      description: The user does not have the permission to perform this action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'
    Conflict:
      description: The request conflicts with current state of the target resource
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/BasicError'